Move tests for queryable and enumerable out of expressions.
Fixes #7904

Refactor `MatchSequencePattern` test to not use `Console.WriteLine()`
and remove it from Linq.Expressions.Tests putting a copy in the tests
for Linq and Linq.Queryable to catch regressions in either.

Simplify `QueryableOfQueryable` and move it into queryable tests.

Remove tests from Expressions test testing features of `Enumerable` or
`Queryable` and already covered by tests in the relevant project.
