+++ src/System.Linq.Expressions/tests/SequenceTests/SequenceTests.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -11,631 +10,6 @@ using Xunit;
 
 namespace System.Linq.Expressions.Tests
 {
-    public class Sequence_Tests
-    {
-        [Fact]
-        public static void QueryableOfQueryable()
-        {
-            IQueryable<int> queryable1 = Queryable.AsQueryable<int>(new int[] { 1, 2, 3 });
-            IQueryable<int>[] queryableArray1 = new IQueryable<int>[] { queryable1, queryable1 };
-            IQueryable<IQueryable<int>> queryable2 = Queryable.AsQueryable<IQueryable<int>>(queryableArray1);
-            ParameterExpression expression1 = Expression.Parameter(typeof(IQueryable<int>), "i");
-            ParameterExpression[] expressionArray1 = new ParameterExpression[] { expression1 };
-            IQueryable<IQueryable<int>> queryable3 = Queryable.Select<IQueryable<int>, IQueryable<int>>((IQueryable<IQueryable<int>>)queryable2, Expression.Lambda<Func<IQueryable<int>, IQueryable<int>>>(expression1, expressionArray1));
-            int i = Queryable.Count<IQueryable<int>>((IQueryable<IQueryable<int>>)queryable3);
-            Assert.Equal(2, i);
-        }
-
-        public struct CustomerRec
-        {
-            public string name;
-            public int?[] total;
-            public CustomerRec(string name, int?[] total) { this.name = name; this.total = total; }
-        }
-
-        public static IEnumerable<int?> indexSelector(CustomerRec cr, int index)
-        {
-            if (index == 4) return new int?[] { 23, 45, 61, 89 };
-            else return new int?[] { };
-        }
-
-        [Fact]
-        public static void SelectMany()
-        {
-            CustomerRec[] source = {new CustomerRec{name="Prakash", total=new int?[]{1, 2, 3, 4}},
-                                    new CustomerRec{name="Bob", total=new int?[]{5, 6}},
-                                    new CustomerRec{name="Chris", total=new int?[]{}},
-                                    new CustomerRec{name=null, total=new int?[]{8, 9}},
-                                    new CustomerRec{name="Prakash", total=new int?[]{-10, 100}}
-                                };
-
-            string[] expected = { "23", "45", "61", "89" };
-            Func<CustomerRec, int, IEnumerable<int?>> collectionSelector = indexSelector;
-            Func<CustomerRec, int?, string> resultSelector = (e, f) => f.ToString();
-
-            var actual = source.SelectMany(collectionSelector, resultSelector);
-
-            int i = 0;
-            foreach (string s in actual)
-            {
-                Assert.Equal(s, expected[i]);
-                i++;
-            }
-        }
-
-        [Fact]
-        public static void BasicWhere()
-        {
-            int[] ints = new int[] { 4, 1, 5, 3, 2 };
-            foreach (int i in ints.Where(x => x < 4))
-            {
-                Assert.True(i < 4);
-            }
-        }
-
-        [Fact]
-        public static void BasicGroupBy()
-        {
-            int[] ints = new int[] { 4, 1, 5, 3, 2 };
-
-            var vs = ints.Select(i => new { a = i, b = i / 2 }).GroupBy(g => g.b);
-
-            foreach (var i in vs)
-            {
-                foreach (var j in i)
-                {
-                    Assert.Equal(j.a / 2, i.Key);
-                }
-            }
-        }
-
-        [Fact]
-        public static void GroupBywithResultSelector()
-        {
-            char[] elements = { 'q', 'q', 'q', 'q', 'q' };
-
-            var result = elements.GroupBy((e) => (e), (e, f) => new { Key = e, Element = f });
-
-            foreach (var e in result)
-            {
-                Assert.Equal(5, e.Element.Count());
-                foreach (var ele in e.Element)
-                    Assert.Equal(ele, 'q');
-            }
-        }
-
-        [Fact]
-        public static void BasicAggregate()
-        {
-            int[] ints = new int[] { 4, 1, 5, 3, 2 };
-
-            // Multiplies all the numbers in ints[].
-            Assert.Equal(120, ints.Aggregate((current, i) => current * i));
-        }
-
-        private delegate bool Predicate<T0>(T0 p0);
-        private delegate bool Predicate<T0, T1>(T0 p0, T1 p1);
-
-        [Fact]
-        public static void DistinctWithNulls()
-        {
-            int?[] ints = new int?[] { 1, 2, null, 4, 2, 1 };
-            var distinct = ints.Distinct().ToList();
-            Assert.Equal(4, distinct.Count);
-            Assert.Equal(1, distinct[0]);
-            Assert.Equal(2, distinct[1]);
-            Assert.Null(distinct[2]);
-            Assert.Equal(4, distinct[3]);
-        }
-
-        [Fact]
-        public static void UnionWithNulls()
-        {
-            int?[] a = new int?[] { 1, 2, null, 4, 2, null, 1 };
-            int?[] b = new int?[] { 2, 4, null, 5, null };
-            var c = a.Union(b).ToList();
-            Assert.Equal(5, c.Count);
-            Assert.Equal(1, c[0]);
-            Assert.Equal(2, c[1]);
-            Assert.Null(c[2]);
-            Assert.Equal(4, c[3]);
-            Assert.Equal(5, c[4]);
-        }
-
-        [Fact]
-        public static void IntersectWithNulls()
-        {
-            int?[] a = new int?[] { 1, 2, null, 4, 2, null, 1 };
-            int?[] b = new int?[] { 2, 4, null, 5, null };
-            var c = a.Intersect(b).ToList();
-            Assert.Equal(3, c.Count);
-            Assert.Equal(2, c[0]);
-            Assert.Null(c[1]);
-            Assert.Equal(4, c[2]);
-        }
-
-        [Fact]
-        public static void ExceptWithNulls()
-        {
-            int?[] a = new int?[] { 1, 2, null, 4, 2, null, 1 };
-            int?[] b = new int?[] { 2, 5 };
-            var c = a.Except(b).ToList();
-            Assert.Equal(3, c.Count);
-            int?[] b2 = new int?[] { 2, 5, null };
-            var c2 = a.Except(b2).ToList();
-            Assert.Equal(2, c2.Count);
-        }
-
-        [Fact]
-        public static void Sequence_NonList()
-        {
-            Assert.Equal(6, Enumerable.Range(0, 10).ElementAtOrDefault(6));
-            Assert.Equal(default(int), Enumerable.Range(0, 10).ElementAtOrDefault(10));
-            Assert.Equal(default(int), Enumerable.Range(0, 10).ElementAtOrDefault(-1));
-
-            Assert.Equal(9, Enumerable.Range(0, 10).Last());
-            Assert.Equal(9, Enumerable.Range(0, 10).LastOrDefault());
-            Assert.Equal(default(int), Enumerable.Range(0, 0).LastOrDefault());
-        }
-
-        [Fact]
-        public static void Contains_SourceNull()
-        {
-            IEnumerable<string> source = null;
-            Assert.Throws<ArgumentNullException>("source", () => source.Contains(null));
-        }
-
-        public static IEnumerable<string> Contains_Source1()
-        {
-            var strings = new[] { "one", "two", null, "four" };
-
-            foreach (var item in strings)
-            {
-                yield return item;
-            }
-        }
-
-        public static IEnumerable<string> Contains_Source2()
-        {
-            var strings = new[] { "one", "two", "three", "four" };
-
-            foreach (var item in strings)
-            {
-                yield return item;
-            }
-        }
-
-        [Fact]
-        public static void Contains_SourceNotCollection()
-        {
-            Assert.True(Contains_Source1().Contains(null));
-            Assert.False(Contains_Source2().Contains(null));
-
-            Assert.False(Contains_Source1().Contains("three"));
-            Assert.True(Contains_Source2().Contains("three"));
-        }
-
-        public static IEnumerable<System.Int16> MaxMin_NumberSourceInt16()
-        {
-            for (System.Int16 i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Int32> MaxMin_NumberSourceInt32()
-        {
-            for (System.Int32 i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Int64> MaxMin_NumberSourceInt64()
-        {
-            for (System.Int64 i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.UInt16> MaxMin_NumberSourceUInt16()
-        {
-            for (System.UInt16 i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.UInt32> MaxMin_NumberSourceUInt32()
-        {
-            for (System.UInt32 i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.UInt64> MaxMin_NumberSourceUInt64()
-        {
-            for (System.UInt64 i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Single> MaxMin_NumberSourceSingle()
-        {
-            for (System.Single i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Double> MaxMin_NumberSourceDouble()
-        {
-            for (System.Double i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Decimal> MaxMin_NumberSourceDecimal()
-        {
-            for (System.Decimal i = 0; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Int16?> MaxMin_NullableNumberSourceInt16()
-        {
-            for (System.Int16 i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.Int16 i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Int32?> MaxMin_NullableNumberSourceInt32()
-        {
-            for (System.Int32 i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.Int32 i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Int64?> MaxMin_NullableNumberSourceInt64()
-        {
-            for (System.Int64 i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.Int64 i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.UInt16?> MaxMin_NullableNumberSourceUInt16()
-        {
-            for (System.UInt16 i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.UInt16 i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.UInt32?> MaxMin_NullableNumberSourceUInt32()
-        {
-            for (System.UInt32 i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.UInt32 i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.UInt64?> MaxMin_NullableNumberSourceUInt64()
-        {
-            for (System.UInt64 i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.UInt64 i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Single?> MaxMin_NullableNumberSourceSingle()
-        {
-            for (System.Single i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.Single i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Double?> MaxMin_NullableNumberSourceDouble()
-        {
-            for (System.Double i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.Double i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Decimal?> MaxMin_NullableNumberSourceDecimal()
-        {
-            for (System.Decimal i = 0; i < 5; i++)
-            {
-                yield return i;
-            }
-
-            yield return null;
-
-            for (System.Decimal i = 5; i < 10; i++)
-            {
-                yield return i;
-            }
-        }
-
-        public static IEnumerable<System.Int16?> MaxMin_AllNull()
-        {
-            for (System.Int16 i = 0; i < 10; i++)
-            {
-                yield return null;
-            }
-        }
-
-        [Fact]
-        public static void Range()
-        {
-            Assert.Equal(0, Enumerable.Range(0, 10).Min());
-            Assert.Equal(9, Enumerable.Range(0, 10).Max());
-            Assert.Equal(10, Enumerable.Range(0, 10).Count());
-            Assert.Equal(0, Enumerable.Range(0, 0).Count());
-        }
-
-        [Fact]
-        public static void Empty()
-        {
-            Assert.Equal(0, Enumerable.Empty<int>().Count());
-        }
-
-        [Fact]
-        public static void RangeNegativeCount()
-        {
-            Assert.Throws<ArgumentOutOfRangeException>("count", () => Enumerable.Range(0, -1));
-        }
-
-        [Fact]
-        public static void MaxEmpty()
-        {
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<int>().Max());
-        }
-
-        [Fact]
-        public static void MinEmpty()
-        {
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<int>().Min());
-        }
-
-        [Fact]
-        public static void MaxNullSource()
-        {
-            Assert.Throws<ArgumentNullException>("source", () => ((IEnumerable<int>)null).Max());
-        }
-
-        [Fact]
-        public static void MinNullSource()
-        {
-            Assert.Throws<ArgumentNullException>("source", () => ((IEnumerable<int>)null).Min());
-        }
-
-        [Fact]
-        public static void MaxMin()
-        {
-            Assert.Null(MaxMin_AllNull().Cast<System.Int16?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.Int32?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.Int64?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt16?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt32?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt64?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.Single?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.Double?>().Max());
-            Assert.Null(MaxMin_AllNull().Cast<System.Decimal?>().Max());
-
-            Assert.Null(MaxMin_AllNull().Cast<System.Int16?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.Int32?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.Int64?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt16?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt32?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt64?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.Single?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.Double?>().Min());
-            Assert.Null(MaxMin_AllNull().Cast<System.Decimal?>().Min());
-
-
-            Assert.Equal(9, MaxMin_NumberSourceInt16().Max());
-            Assert.Equal(9, MaxMin_NumberSourceInt32().Max());
-            Assert.Equal(9, MaxMin_NumberSourceInt64().Max());
-            Assert.Equal(9, MaxMin_NumberSourceUInt16().Max());
-            Assert.Equal((uint)9, MaxMin_NumberSourceUInt32().Max());
-            Assert.Equal((ulong)9, MaxMin_NumberSourceUInt64().Max());
-            Assert.Equal(9, MaxMin_NumberSourceSingle().Max());
-            Assert.Equal(9, MaxMin_NumberSourceDouble().Max());
-            Assert.Equal(9, MaxMin_NumberSourceDecimal().Max());
-
-            Assert.Equal(0, MaxMin_NumberSourceInt16().Min());
-            Assert.Equal(0, MaxMin_NumberSourceInt32().Min());
-            Assert.Equal(0, MaxMin_NumberSourceInt64().Min());
-            Assert.Equal(0, MaxMin_NumberSourceUInt16().Min());
-            Assert.Equal((uint)0, MaxMin_NumberSourceUInt32().Min());
-            Assert.Equal((ulong)0, MaxMin_NumberSourceUInt64().Min());
-            Assert.Equal(0, MaxMin_NumberSourceSingle().Min());
-            Assert.Equal(0, MaxMin_NumberSourceDouble().Min());
-            Assert.Equal(0, MaxMin_NumberSourceDecimal().Min());
-
-
-            Assert.Equal((short?)9, MaxMin_NullableNumberSourceInt16().Max());
-            Assert.Equal(9, MaxMin_NullableNumberSourceInt32().Max());
-            Assert.Equal(9, MaxMin_NullableNumberSourceInt64().Max());
-            Assert.Equal((ushort?)9, MaxMin_NullableNumberSourceUInt16().Max());
-            Assert.Equal((uint?)9, MaxMin_NullableNumberSourceUInt32().Max());
-            Assert.Equal((ulong?)9, MaxMin_NullableNumberSourceUInt64().Max());
-            Assert.Equal(9, MaxMin_NullableNumberSourceSingle().Max());
-            Assert.Equal(9, MaxMin_NullableNumberSourceDouble().Max());
-            Assert.Equal(9, MaxMin_NullableNumberSourceDecimal().Max());
-
-            Assert.Equal((short?)0, MaxMin_NullableNumberSourceInt16().Min());
-            Assert.Equal(0, MaxMin_NullableNumberSourceInt32().Min());
-            Assert.Equal(0, MaxMin_NullableNumberSourceInt64().Min());
-            Assert.Equal((ushort?)0, MaxMin_NullableNumberSourceUInt16().Min());
-            Assert.Equal((uint?)0, MaxMin_NullableNumberSourceUInt32().Min());
-            Assert.Equal((ulong?)0, MaxMin_NullableNumberSourceUInt64().Min());
-            Assert.Equal(0, MaxMin_NullableNumberSourceSingle().Min());
-            Assert.Equal(0, MaxMin_NullableNumberSourceDouble().Min());
-            Assert.Equal(0, MaxMin_NullableNumberSourceDecimal().Min());
-
-
-            Assert.Null(MaxMin_AllNull().Cast<System.Int16?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Int32?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Int64?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt16?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt32?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt64?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Single?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Double?>().Max(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Decimal?>().Max(item => item));
-
-            Assert.Null(MaxMin_AllNull().Cast<System.Int16?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Int32?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Int64?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt16?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt32?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.UInt64?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Single?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Double?>().Min(item => item));
-            Assert.Null(MaxMin_AllNull().Cast<System.Decimal?>().Min(item => item));
-
-
-            Assert.Equal(9, MaxMin_NumberSourceInt16().Max(item => item));
-            Assert.Equal(9, MaxMin_NumberSourceInt32().Max(item => item));
-            Assert.Equal(9, MaxMin_NumberSourceInt64().Max(item => item));
-            Assert.Equal(9, MaxMin_NumberSourceUInt16().Max(item => item));
-            Assert.Equal((uint)9, MaxMin_NumberSourceUInt32().Max(item => item));
-            Assert.Equal((ulong)9, MaxMin_NumberSourceUInt64().Max(item => item));
-            Assert.Equal(9, MaxMin_NumberSourceSingle().Max(item => item));
-            Assert.Equal(9, MaxMin_NumberSourceDouble().Max(item => item));
-            Assert.Equal(9, MaxMin_NumberSourceDecimal().Max(item => item));
-
-            Assert.Equal(0, MaxMin_NumberSourceInt16().Min(item => item));
-            Assert.Equal(0, MaxMin_NumberSourceInt32().Min(item => item));
-            Assert.Equal(0, MaxMin_NumberSourceInt64().Min(item => item));
-            Assert.Equal(0, MaxMin_NumberSourceUInt16().Min(item => item));
-            Assert.Equal((uint)0, MaxMin_NumberSourceUInt32().Min(item => item));
-            Assert.Equal((ulong)0, MaxMin_NumberSourceUInt64().Min(item => item));
-            Assert.Equal(0, MaxMin_NumberSourceSingle().Min(item => item));
-            Assert.Equal(0, MaxMin_NumberSourceDouble().Min(item => item));
-            Assert.Equal(0, MaxMin_NumberSourceDecimal().Min(item => item));
-
-
-            Assert.Equal((short?)9, MaxMin_NullableNumberSourceInt16().Max(item => item));
-            Assert.Equal(9, MaxMin_NullableNumberSourceInt32().Max(item => item));
-            Assert.Equal(9, MaxMin_NullableNumberSourceInt64().Max(item => item));
-            Assert.Equal((ushort?)9, MaxMin_NullableNumberSourceUInt16().Max(item => item));
-            Assert.Equal((ushort?)9, MaxMin_NullableNumberSourceUInt32().Max(item => item));
-            Assert.Equal((ulong?)9, MaxMin_NullableNumberSourceUInt64().Max(item => item));
-            Assert.Equal(9, MaxMin_NullableNumberSourceSingle().Max(item => item));
-            Assert.Equal(9, MaxMin_NullableNumberSourceDouble().Max(item => item));
-            Assert.Equal(9, MaxMin_NullableNumberSourceDecimal().Max(item => item));
-
-            Assert.Equal((short?)0, MaxMin_NullableNumberSourceInt16().Min(item => item));
-            Assert.Equal(0, MaxMin_NullableNumberSourceInt32().Min(item => item));
-            Assert.Equal(0, MaxMin_NullableNumberSourceInt64().Min(item => item));
-            Assert.Equal((ushort?)0, MaxMin_NullableNumberSourceUInt16().Min(item => item));
-            Assert.Equal((uint?)0, MaxMin_NullableNumberSourceUInt32().Min(item => item));
-            Assert.Equal((ulong?)0, MaxMin_NullableNumberSourceUInt64().Min(item => item));
-            Assert.Equal(0, MaxMin_NullableNumberSourceSingle().Min(item => item));
-            Assert.Equal(0, MaxMin_NullableNumberSourceDouble().Min(item => item));
-            Assert.Equal(0, MaxMin_NullableNumberSourceDecimal().Min(item => item));
-        }
-
-        [Fact]
-        public static void Lookup()
-        {
-            Assert.Equal(9, Enumerable.Range(0, 10).ToDictionary(t => t.ToString())["9"]);
-
-            ILookup<string, int> lookup = Enumerable.Range(0, 10).ToLookup(t => t.ToString(), t => t, EqualityComparer<string>.Default);
-
-            Assert.Equal(1, ((ICollection<int>)lookup["9"]).Count);
-
-            Assert.Equal(1, ((ICollection<int>)lookup["9"]).Count);
-
-            Assert.Equal(9, ((IList<int>)lookup["9"])[0]);
-
-            Assert.False(((ICollection<int>)lookup["9"]).Contains(3));
-
-            Assert.True(((ICollection<int>)lookup["9"]).IsReadOnly);
-        }
-
-        [Fact]
-        public static void GroupingKeyIsPublic()
-        {
-            // Grouping.Key needs to be public (not explicitly implemented) for the sake of WPF.
-
-            object[] objs = { "Foo", 1.0M, "Bar", new { X = "X" }, 2.00M };
-            object group = objs.GroupBy(x => x.GetType()).First();
-
-            Type grouptype = group.GetType();
-            PropertyInfo key = grouptype.GetProperty("Key", BindingFlags.Instance | BindingFlags.Public);
-            Assert.NotNull(key);
-        }
-    }
-
     public class Expression_Tests
     {
         [Fact]
@@ -1237,531 +611,6 @@ namespace System.Linq.Expressions.Tests
             Assert.Equal(name, mc.Method.Name);
 
             Assert.Equal(typeof(Queryable), mc.Method.DeclaringType);
-        }
-    }
-
-    public class Queryable_Tests
-    {
-        [Fact]
-        public static void SequenceQueryToString()
-        {
-            var q = (new int[] { 1, 2, 3, 4, 5, 6 }).AsQueryable();
-            var s = q.ToString();
-        }
-
-        [Theory]
-        [ClassData(typeof(CompilationTypes))]
-        public static void DynamicSequenceQuery(bool useInterpreter)
-        {
-            ParameterExpression i = Expression.Parameter(typeof(int[]), "i");
-            var e = Expression.Lambda<Func<int[], IEnumerable<int>>>(
-                  Expression.Convert(
-                    Expression.Call(
-                       null,
-                       typeof(Queryable).GetMethod("AsQueryable", new[] { typeof(IEnumerable) }),
-                       new Expression[] {
-                        Expression.Convert(i, typeof(IEnumerable<int>))
-                        }
-                       ),
-                    typeof(IEnumerable<int>)
-                    ),
-                    new ParameterExpression[] { i }
-                  );
-
-            string result = "";
-            foreach (var x in e.Compile(useInterpreter)(new[] { 1, 2, 3 }))
-                result += ", " + x.ToString();
-
-            Assert.Equal(", 1, 2, 3", result);
-        }
-
-        [Fact(Skip = "870811")]
-        public static void QueryableMethods()
-        {
-            var s = new int[] { 1, 2, 3, 4, 5, 6 };
-            var q = s.AsQueryable();
-            var s2 = new int[] { 3, 5 };
-            var q2 = s2.AsQueryable();
-            var s3 = new int[] { -1, -2, 3, 4 };
-            var q3 = s3.AsQueryable();
-            var s4 = new int[] { -1, -2 };
-            var q4 = s4.AsQueryable();
-            var snInt = new int?[] { 1, 2, null, 4, 5, 6 };
-            var qnInt = snInt.AsQueryable();
-            var snDouble = new double?[] { 1, 2, null, 4, 5, 6 };
-            var qnDouble = snDouble.AsQueryable();
-            var snLong = new long?[] { 1, 2, null, 4, 5, 6 };
-            var qnLong = snLong.AsQueryable();
-            var snDecimal = new decimal?[] { 1, 2, null, 4, 5, 6 };
-            var qnDecimal = snDecimal.AsQueryable();
-            var sInt = new int[] { 1, 2, 3, 4, 5, 6 };
-            var qInt = sInt.AsQueryable();
-            var sDouble = new double[] { 1, 2, 3, 4, 5, 6 };
-            var qDouble = sDouble.AsQueryable();
-            var sLong = new long[] { 1, 2, 3, 4, 5, 6 };
-            var qLong = sLong.AsQueryable();
-            var sDecimal = new decimal[] { 1, 2, 3, 4, 5, 6 };
-            var qDecimal = sDecimal.AsQueryable();
-            var se = new int[] { };
-            var qe = se.AsQueryable();
-            var ss = new int[] { 9 };
-            var qs = ss.AsQueryable();
-            string[] names = new string[] { "zero", "one", "two", "three", "four", "five", "six" };
-            var comp = StringComparer.OrdinalIgnoreCase;
-
-            Assert.Equal(21, q.Aggregate((a, b) => a + b));
-            Assert.Equal(21, q.Aggregate(0, (a, b) => a + b));
-            Assert.Equal("21", q.Aggregate(0, (a, b) => a + b, (r) => r.ToString()));
-            Assert.True(q.All(x => x < 7));
-            Assert.False(q.All(x => x < 4));
-            Assert.True(q.Any());
-            Assert.False(qe.Any());
-            Assert.True(q.Any(x => x > 5));
-            Assert.False(q.Any(x => x < 0));
-            Assert.Equal(6, q.Cast<object>().Count());
-            Assert.Equal(12, q.Concat(q).Count());
-            Assert.Equal(12, s.Concat(q).Count());
-            Assert.Equal(12, q.Concat(s).Count());
-            Assert.Equal(3, q.Count(x => x < 4));
-            Assert.True(q.Contains(3));
-            Assert.False(q.Contains(9));
-            Assert.False(qe.Contains(5));
-            Assert.Equal(6, q.Distinct().Count());
-            Assert.Equal(1, qe.DefaultIfEmpty().Count());
-            Assert.Equal(4, q.ElementAt(3));
-            Assert.Equal(4, q.ToList().ElementAt(3));
-            Assert.Equal(4, s.ElementAt(3));
-            Assert.Equal(4, s.ToList().ElementAt(3));
-            Assert.Equal(0, q.ElementAtOrDefault(10));
-            Assert.Equal(0, q.ToList().ElementAtOrDefault(10));
-            Assert.Equal(4, s.ElementAtOrDefault(3));
-            Assert.Equal(4, s.ToList().ElementAtOrDefault(3));
-            Assert.Equal(0, s.ElementAtOrDefault(10));
-            Assert.Equal(0, s.ToList().ElementAtOrDefault(10));
-            Assert.True(q.SequenceEqual(((IEnumerable<int>)q).AsEnumerable()));
-            Assert.True(q.SequenceEqual(q));
-            Assert.True(s.SequenceEqual(q));
-            Assert.True(q.SequenceEqual(s));
-            Assert.False(q.SequenceEqual(q2));
-            Assert.False(q.SequenceEqual(q2.Reverse()));
-            Assert.False(q2.SequenceEqual(q));
-            Assert.False(q2.SequenceEqual(q.Reverse()));
-            Assert.False(q3.SequenceEqual(q4));
-            Assert.False(q4.SequenceEqual(q3));
-            Assert.Equal(4, q.Except(q2).Count());
-            Assert.Equal(4, s.Except(q2).Count());
-            Assert.Equal(4, q.Except(s2).Count());
-            Assert.Equal(1, q.First());
-            Assert.Equal(4, q.First(x => x > 3));
-            Assert.Equal(4, ((IEnumerable<int>)q).First(x => x > 3));
-            Assert.Equal(0, qe.FirstOrDefault());
-            Assert.Equal(1, q.FirstOrDefault());
-            Assert.Equal(1, q.ToList().FirstOrDefault());
-            Assert.Equal(1, Enumerable.Range(1, 6).FirstOrDefault());
-            Assert.Equal(1, Enumerable.Range(1, 6).ToList().FirstOrDefault());
-            Assert.Equal(6, q.FirstOrDefault(x => x > 5));
-            Assert.Equal(0, q.FirstOrDefault(x => x > 10));
-            Assert.Equal(6, q.ToList().FirstOrDefault(x => x > 5));
-            Assert.Equal(0, q.ToList().FirstOrDefault(x => x > 10));
-            Assert.Equal(6, Enumerable.Range(1, 6).FirstOrDefault(x => x > 5));
-            Assert.Equal(0, Enumerable.Range(1, 6).FirstOrDefault(x => x > 10));
-            Assert.Equal(4, q.GroupBy(x => x / 2).Count());
-            Assert.Equal(3, q.GroupBy(x => names[x].Length, x => x).Count());
-            Assert.Equal(6, q.GroupBy(x => names[x]).Count());
-            Assert.Equal(6, q.GroupBy(x => names[x], x => x, comp).Count());
-            Assert.Equal(2, q.Intersect(q2).Count());
-            Assert.Equal(6, q.Last());
-            Assert.Equal(6, q.Last(x => x > 3));
-            Assert.Equal(0, qe.LastOrDefault());
-            Assert.Equal(6, s.LastOrDefault());
-            Assert.Equal(6, q.LastOrDefault());
-            Assert.Equal(6, q.ToList().LastOrDefault());
-            Assert.Equal(6, Enumerable.Range(1, 6).LastOrDefault());
-            Assert.Equal(6, q.LastOrDefault(x => x > 5));
-            Assert.Equal(0, q.LastOrDefault(x => x > 10));
-            Assert.Equal(6, q.ToList().LastOrDefault(x => x > 5));
-            Assert.Equal(0, q.ToList().LastOrDefault(x => x > 10));
-            Assert.Equal(6, Enumerable.Range(1, 6).ToList().LastOrDefault(x => x > 5));
-            Assert.Equal(0, Enumerable.Range(1, 6).ToList().LastOrDefault(x => x > 10));
-            Assert.Equal(6L, q.LongCount());
-            Assert.Equal(3L, q.LongCount(x => x <= 3));
-            Assert.Equal(6, q.Max());
-            Assert.Equal(36, q.Max(x => x * x));
-            Assert.Equal(6, q.Reverse().Max());
-            Assert.Equal(36, q.Reverse().Max(x => x * x));
-            Assert.Equal(1, q.Min());
-            Assert.Equal(1, q.Min(x => x * x));
-            Assert.Equal(1, q.Reverse().Min());
-            Assert.Equal(1, q.Reverse().Min(x => x * x));
-
-            Assert.Equal(6, sInt.Max());
-            Assert.Equal(36, sInt.Max(x => x * x));
-            Assert.Equal(6, sInt.Reverse().Max());
-            Assert.Equal(36, sInt.Reverse().Max(x => x * x));
-            Assert.Equal(1, sInt.Min());
-            Assert.Equal(1, sInt.Min(x => x * x));
-            Assert.Equal(1, sInt.Reverse().Min());
-            Assert.Equal(1, sInt.Reverse().Min(x => x * x));
-            Assert.Equal(6, sDouble.Max());
-            Assert.Equal(36, sDouble.Max(x => x * x));
-            Assert.Equal(6, sDouble.Reverse().Max());
-            Assert.Equal(36, sDouble.Reverse().Max(x => x * x));
-            Assert.Equal(1, sDouble.Min());
-            Assert.Equal(1, sDouble.Min(x => x * x));
-            Assert.Equal(1, sDouble.Reverse().Min());
-            Assert.Equal(1, sDouble.Reverse().Min(x => x * x));
-            Assert.Equal(6, sLong.Max());
-            Assert.Equal(36, sLong.Max(x => x * x));
-            Assert.Equal(6, sLong.Reverse().Max());
-            Assert.Equal(36, sLong.Reverse().Max(x => x * x));
-            Assert.Equal(1, sLong.Min());
-            Assert.Equal(1, sLong.Min(x => x * x));
-            Assert.Equal(1, sLong.Reverse().Min());
-            Assert.Equal(1, sLong.Reverse().Min(x => x * x));
-            Assert.Equal(6, sDecimal.Max());
-            Assert.Equal(36, sDecimal.Max(x => x * x));
-            Assert.Equal(6, sDecimal.Reverse().Max());
-            Assert.Equal(36, sDecimal.Reverse().Max(x => x * x));
-            Assert.Equal(1, sDecimal.Min());
-            Assert.Equal(1, sDecimal.Min(x => x * x));
-            Assert.Equal(1, sDecimal.Reverse().Min());
-            Assert.Equal(1, sDecimal.Reverse().Min(x => x * x));
-
-            Assert.Equal(6, snInt.Max());
-            Assert.Equal(36, snInt.Max(x => x * x));
-            Assert.Equal(6, snInt.Reverse().Max());
-            Assert.Equal(36, snInt.Reverse().Max(x => x * x));
-            Assert.Equal(1, snInt.Min());
-            Assert.Equal(1, snInt.Min(x => x * x));
-            Assert.Equal(1, snInt.Reverse().Min());
-            Assert.Equal(1, snInt.Reverse().Min(x => x * x));
-            Assert.Equal(6, snDouble.Max());
-            Assert.Equal(36, snDouble.Max(x => x * x));
-            Assert.Equal(6, snDouble.Reverse().Max());
-            Assert.Equal(36, snDouble.Reverse().Max(x => x * x));
-            Assert.Equal(1, snDouble.Min());
-            Assert.Equal(1, snDouble.Min(x => x * x));
-            Assert.Equal(1, snDouble.Reverse().Min());
-            Assert.Equal(1, snDouble.Reverse().Min(x => x * x));
-            Assert.Equal(6, snLong.Max());
-            Assert.Equal(36, snLong.Max(x => x * x));
-            Assert.Equal(6, snLong.Reverse().Max());
-            Assert.Equal(36, snLong.Reverse().Max(x => x * x));
-            Assert.Equal(1, snLong.Min());
-            Assert.Equal(1, snLong.Min(x => x * x));
-            Assert.Equal(1, snLong.Reverse().Min());
-            Assert.Equal(1, snLong.Reverse().Min(x => x * x));
-            Assert.Equal(6, snDecimal.Max());
-            Assert.Equal(36, snDecimal.Max(x => x * x));
-            Assert.Equal(6, snDecimal.Reverse().Max());
-            Assert.Equal(36, snDecimal.Reverse().Max(x => x * x));
-            Assert.Equal(1, snDecimal.Min());
-            Assert.Equal(1, snDecimal.Min(x => x * x));
-            Assert.Equal(1, snDecimal.Reverse().Min());
-            Assert.Equal(1, snDecimal.Reverse().Min(x => x * x));
-
-
-            Assert.Equal(6, qInt.Max());
-            Assert.Equal(36, qInt.Max(x => x * x));
-            Assert.Equal(6, qInt.Reverse().Max());
-            Assert.Equal(36, qInt.Reverse().Max(x => x * x));
-            Assert.Equal(1, qInt.Min());
-            Assert.Equal(1, qInt.Min(x => x * x));
-            Assert.Equal(1, qInt.Reverse().Min());
-            Assert.Equal(1, qInt.Reverse().Min(x => x * x));
-            Assert.Equal(6, qDouble.Max());
-            Assert.Equal(36, qDouble.Max(x => x * x));
-            Assert.Equal(6, qDouble.Reverse().Max());
-            Assert.Equal(36, qDouble.Reverse().Max(x => x * x));
-            Assert.Equal(1, qDouble.Min());
-            Assert.Equal(1, qDouble.Min(x => x * x));
-            Assert.Equal(1, qDouble.Reverse().Min());
-            Assert.Equal(1, qDouble.Reverse().Min(x => x * x));
-            Assert.Equal(6, qLong.Max());
-            Assert.Equal(36, qLong.Max(x => x * x));
-            Assert.Equal(6, qLong.Reverse().Max());
-            Assert.Equal(36, qLong.Reverse().Max(x => x * x));
-            Assert.Equal(1, qLong.Min());
-            Assert.Equal(1, qLong.Min(x => x * x));
-            Assert.Equal(1, qLong.Reverse().Min());
-            Assert.Equal(1, qLong.Reverse().Min(x => x * x));
-            Assert.Equal(6, qDecimal.Max());
-            Assert.Equal(36, qDecimal.Max(x => x * x));
-            Assert.Equal(6, qDecimal.Reverse().Max());
-            Assert.Equal(36, qDecimal.Reverse().Max(x => x * x));
-            Assert.Equal(1, qDecimal.Min());
-            Assert.Equal(1, qDecimal.Min(x => x * x));
-            Assert.Equal(1, qDecimal.Reverse().Min());
-            Assert.Equal(1, qDecimal.Reverse().Min(x => x * x));
-            Assert.Equal(6, qnInt.Max());
-            Assert.Equal(36, qnInt.Max(x => x * x));
-            Assert.Equal(6, qnInt.Reverse().Max());
-            Assert.Equal(36, qnInt.Reverse().Max(x => x * x));
-            Assert.Equal(1, qnInt.Min());
-            Assert.Equal(1, qnInt.Min(x => x * x));
-            Assert.Equal(1, qnInt.Reverse().Min());
-            Assert.Equal(1, qnInt.Reverse().Min(x => x * x));
-            Assert.Equal(6, qnDouble.Max());
-            Assert.Equal(36, qnDouble.Max(x => x * x));
-            Assert.Equal(6, qnDouble.Reverse().Max());
-            Assert.Equal(36, qnDouble.Reverse().Max(x => x * x));
-            Assert.Equal(1, qnDouble.Min());
-            Assert.Equal(1, qnDouble.Min(x => x * x));
-            Assert.Equal(1, qnDouble.Reverse().Min());
-            Assert.Equal(1, qnDouble.Reverse().Min(x => x * x));
-            Assert.Equal(6, qnLong.Max());
-            Assert.Equal(36, qnLong.Max(x => x * x));
-            Assert.Equal(6, qnLong.Reverse().Max());
-            Assert.Equal(36, qnLong.Reverse().Max(x => x * x));
-            Assert.Equal(1, qnLong.Min());
-            Assert.Equal(1, qnLong.Min(x => x * x));
-            Assert.Equal(1, qnLong.Reverse().Min());
-            Assert.Equal(1, qnLong.Reverse().Min(x => x * x));
-            Assert.Equal(6, qnDecimal.Max());
-            Assert.Equal(36, qnDecimal.Max(x => x * x));
-            Assert.Equal(6, qnDecimal.Reverse().Max());
-            Assert.Equal(36, qnDecimal.Reverse().Max(x => x * x));
-            Assert.Equal(1, qnDecimal.Min());
-            Assert.Equal(1, qnDecimal.Min(x => x * x));
-            Assert.Equal(1, qnDecimal.Reverse().Min());
-            Assert.Equal(1, qnDecimal.Reverse().Min(x => x * x));
-            Assert.Equal(6, q.OfType<int>().Count());
-            Assert.Equal(0, q.OfType<string>().Count());
-            Assert.Equal(5, q.OrderBy(x => names[x]).First());
-            Assert.Equal(2, q.OrderByDescending(x => names[x]).First());
-            Assert.Equal(2, q.OrderBy(x => names[x / 2]).ThenBy(x => x).First());
-            Assert.Equal(1, q.OrderByDescending(x => names[x / 2]).ThenByDescending(x => x).First());
-            Assert.Equal(5, q.OrderBy(x => names[x], comp).First());
-            Assert.Equal(2, q.OrderByDescending(x => names[x], comp).First());
-            Assert.Equal(2, q.OrderBy(x => names[x / 2], comp).ThenBy(x => x).First());
-            Assert.Equal(1, q.OrderByDescending(x => names[x / 2], comp).ThenByDescending(x => x).First());
-            Assert.Equal(6, q.Select(x => x * x).Count());
-            Assert.Equal(12, q.SelectMany(x => q2).Count());
-            Assert.Equal(12, q.SelectMany(x => q2, (x, y) => new { x, y }).Count());
-            Assert.Equal(6, q.Select((x, i) => x * i).Count());
-            Assert.Equal(12, q.SelectMany((x, i) => q2).Count());
-            Assert.Equal(9, qs.Single());
-            Assert.Equal(6, q.Single(x => x > 5));
-            Assert.Equal(9, qs.SingleOrDefault());
-            Assert.Equal(0, qe.SingleOrDefault());
-            Assert.Equal(6, q.SingleOrDefault(x => x > 5));
-            Assert.Equal(0, q.SingleOrDefault(x => x > 10));
-            Assert.Equal(4, q.Skip(2).Count());
-            Assert.Equal(3, q.SkipWhile(x => x < 4).Count());
-            Assert.Equal(3, q.SkipWhile((x, i) => x < 4).Count());
-            Assert.Equal(3, q.Take(3).Count());
-            Assert.Equal(3, q.TakeWhile(x => x < 4).Count());
-            Assert.Equal(3, q.TakeWhile((x, i) => x < 4).Count());
-            Assert.Equal(6, q.Union(q).Count());
-            Assert.Equal(6, s.Union(q).Count());
-            Assert.Equal(6, q.Union(s).Count());
-            Assert.Equal(7, q.Union(qs).Count());
-            Assert.Equal(3, q.Where(x => x > 3).Count());
-            Assert.Equal(6, q.Where((x, i) => x > i).Count());
-
-            Assert.Equal(2, q.Join(q2, v => v, v => v, (x, y) => new { x, y }).Count());
-            Assert.Equal(2, s.Join(q2, v => v, v => v, (x, y) => new { x, y }).Count());
-            Assert.Equal(2, q.Join(s2, v => v, v => v, (x, y) => new { x, y }).Count());
-            Assert.Equal(6, q.GroupJoin(q2, v => v, v => v, (x, ys) => new { x, ys }).Count());
-            Assert.Equal(6, s.GroupJoin(q2, v => v, v => v, (x, ys) => new { x, ys }).Count());
-            Assert.Equal(6, q.GroupJoin(s2, v => v, v => v, (x, ys) => new { x, ys }).Count());
-
-            // Aggregates  (count and longcount included above)
-            Assert.Equal(1, q.Min());
-            Assert.Equal(1, q.Min(x => x));
-            Assert.Equal(6, q.Max());
-            Assert.Equal(6, q.Max(x => x));
-            Assert.Equal(21, q.Sum());
-            Assert.Equal(21, q.Select(x => (int?)x).Sum());
-            Assert.Equal(21, q.Select(x => (long)x).Sum());
-            Assert.Equal(21, q.Select(x => (long?)x).Sum());
-            Assert.Equal(21, q.Select(x => (double)x).Sum());
-            Assert.Equal(21, q.Select(x => (double?)x).Sum());
-            Assert.Equal(21, q.Select(x => (decimal)x).Sum());
-            Assert.Equal(21, q.Select(x => (decimal?)x).Sum());
-            Assert.Equal(21, q.Sum(x => x));
-            Assert.Equal(21, q.Sum(x => (int?)x));
-            Assert.Equal(21, q.Sum(x => (long)x));
-            Assert.Equal(21, q.Sum(x => (long?)x));
-            Assert.Equal(21, q.Sum(x => (double)x));
-            Assert.Equal(21, q.Sum(x => (double?)x));
-            Assert.Equal(21, q.Sum(x => (decimal)x));
-            Assert.Equal(21, q.Sum(x => (decimal?)x));
-            Assert.Equal(3.5, q.Average());
-            Assert.Equal(3.5, q.Select(x => (int?)x).Average());
-            Assert.Equal(3.5, q.Select(x => (long)x).Average());
-            Assert.Equal(3.5, q.Select(x => (long?)x).Average());
-            Assert.Equal(3.5, q.Select(x => (double)x).Average());
-            Assert.Equal(3.5, q.Select(x => (double?)x).Average());
-            Assert.Equal(3.5m, q.Select(x => (decimal)x).Average());
-            Assert.Equal(3.5m, q.Select(x => (decimal?)x).Average());
-            Assert.Equal(3.5, q.Average(x => x));
-            Assert.Equal(3.5, q.Average(x => (int?)x));
-            Assert.Equal(3.5, q.Average(x => (long)x));
-            Assert.Equal(3.5, q.Average(x => (long?)x));
-            Assert.Equal(3.5, q.Average(x => (double)x));
-            Assert.Equal(3.5, q.Average(x => (double?)x));
-            Assert.Equal(3.5m, q.Average(x => (decimal)x));
-            Assert.Equal(3.5m, q.Average(x => (decimal?)x));
-            Assert.Null((new int?[] { }).Average());
-            Assert.Null((new long?[] { }).Average());
-            Assert.Null((new double?[] { }).Average());
-            Assert.Null((new decimal?[] { }).Average());
-
-            Assert.Equal(13, s.Zip(s2, (a, b) => a * b).Sum());
-            Assert.Equal(13, q.Zip(q2, (a, b) => a * b).Sum());
-            Assert.Equal(34, s2.Zip(s2, (a, b) => a * b).Sum());
-            Assert.Equal(34, q2.Zip(q2, (a, b) => a * b).Sum());
-        }
-
-        [Fact]
-        public static void MatchSequencePattern()
-        {
-            List<MethodInfo> list = GetMissingExtensionMethods(
-                typeof(System.Linq.Enumerable),
-                typeof(System.Linq.Queryable),
-                 new string[] {
-                     "ToLookup",
-                     "ToDictionary",
-                     "ToArray",
-                     "AsEnumerable",
-                     "ToList",
-                     "Fold",
-                     "LeftJoin",
-                     "Append",
-                     "Prepend",
-                 }
-                ).ToList();
-
-            if (list.Count > 0)
-            {
-                Console.WriteLine("Enumerable methods not defined by Queryable\n");
-                foreach (MethodInfo m in list)
-                {
-                    Console.WriteLine(m);
-                }
-                Console.WriteLine();
-            }
-
-            List<MethodInfo> list2 = GetMissingExtensionMethods(
-                typeof(System.Linq.Queryable),
-                typeof(System.Linq.Enumerable),
-                 new string[] {
-                     "AsQueryable"
-                 }
-                ).ToList();
-
-            if (list2.Count > 0)
-            {
-                Console.WriteLine("Queryable methods not defined by Enumerable\n");
-                foreach (MethodInfo m in list2)
-                {
-                    Console.WriteLine(m);
-                }
-            }
-            Assert.Equal(Enumerable.Empty<string>(), list.Select(mi => mi.Name));
-            Assert.Equal(Enumerable.Empty<string>(), list2.Select(mi => mi.Name));
-            Assert.True(list.Count == 0 && list2.Count == 0);
-        }
-
-        private static IEnumerable<MethodInfo> GetMissingExtensionMethods(Type a, Type b, string[] excludedMethods)
-        {
-            var dex = excludedMethods.ToDictionary(s => s);
-
-            var aMethods =
-                a.GetMethods(BindingFlags.Static | BindingFlags.Public)
-                .Where(m => m.CustomAttributes.Any(c => c.AttributeType == typeof(System.Runtime.CompilerServices.ExtensionAttribute)))
-                .ToLookup(m => m.Name);
-
-            MethodComparer mc = new MethodComparer();
-            var bMethods = b.GetMethods(BindingFlags.Static | BindingFlags.Public)
-                .Where(m => m.CustomAttributes.Any(c => c.AttributeType == typeof(System.Runtime.CompilerServices.ExtensionAttribute)))
-                .ToLookup(m => m, mc);
-
-            // look for sequence operator methods not defined by DLINQ
-            foreach (var group in aMethods)
-            {
-                if (dex.ContainsKey(group.Key))
-                    continue;
-                foreach (MethodInfo m in group)
-                {
-                    if (!bMethods.Contains(m))
-                        yield return m;
-                }
-            }
-        }
-
-        private class MethodComparer : IEqualityComparer<MethodInfo>
-        {
-            public int GetHashCode(MethodInfo m)
-            {
-                return m.Name.GetHashCode();
-            }
-
-            public bool Equals(MethodInfo a, MethodInfo b)
-            {
-                if (a.Name != b.Name)
-                    return false;
-                ParameterInfo[] pas = a.GetParameters();
-                ParameterInfo[] pbs = b.GetParameters();
-                if (pas.Length != pbs.Length)
-                    return false;
-                Type[] aArgs = a.GetGenericArguments();
-                Type[] bArgs = b.GetGenericArguments();
-                for (int i = 0, n = pas.Length; i < n; i++)
-                {
-                    ParameterInfo pa = pas[i];
-                    ParameterInfo pb = pbs[i];
-                    Type ta = Strip(pa.ParameterType);
-                    Type tb = Strip(pb.ParameterType);
-                    if (ta.GetTypeInfo().IsGenericType && tb.GetTypeInfo().IsGenericType)
-                    {
-                        if (ta.GetGenericTypeDefinition() != tb.GetGenericTypeDefinition())
-                            return false;
-                    }
-                    else if (ta.IsGenericParameter && tb.IsGenericParameter)
-                    {
-                        return this.GetArgIndex(aArgs, ta) == this.GetArgIndex(bArgs, tb);
-                    }
-                    else if (ta != tb)
-                    {
-                        return false;
-                    }
-                }
-                return true;
-            }
-
-            private int GetArgIndex(Type[] args, Type arg)
-            {
-                for (int i = 0, n = args.Length; i < n; i++)
-                {
-                    if (args[i] == arg)
-                        return i;
-                }
-                return -1;
-            }
-
-            private Type Strip(Type t)
-            {
-                if (t.GetTypeInfo().IsGenericType)
-                {
-                    Type g = t;
-                    if (!g.GetTypeInfo().IsGenericTypeDefinition)
-                        g = t.GetGenericTypeDefinition();
-                    if (g == typeof(IQueryable<>) || g == typeof(IEnumerable<>))
-                        return typeof(IEnumerable);
-                    if (g == typeof(Expression<>))
-                        return t.GetGenericArguments()[0];
-                    if (g == typeof(IOrderedEnumerable<>) || g == typeof(IOrderedQueryable<>))
-                        return typeof(IOrderedQueryable);
-                }
-                else
-                {
-                    if (t == typeof(IQueryable))
-                        return typeof(IEnumerable);
-                }
-                return t;
-            }
         }
     }
