+++ src/System.Linq.Expressions/tests/project.json
@@ -2,7 +2,6 @@
   "dependencies": {
     "Microsoft.NETCore.Platforms": "1.0.1-rc3-24026-00",
     "System.Collections": "4.0.11-rc3-24026-00",
-    "System.Console": "4.0.0-rc3-24026-00",
     "System.Globalization": "4.0.11-rc3-24026-00",
     "System.Linq": "4.1.0-rc3-24026-00",
     "System.Linq.Expressions": "4.1.0-rc3-24026-00",
