+++ src/System.Linq/tests/project.json
@@ -5,7 +5,9 @@
     "System.Collections": "4.0.11-rc3-24026-00",
     "System.Diagnostics.Debug": "4.0.11-rc3-24026-00",
     "System.Linq.Expressions": "4.1.0-rc3-24026-00",
+    "System.Linq.Queryable": "4.0.1-rc3-24026-00",
     "System.ObjectModel": "4.0.12-rc3-24026-00",
+    "System.Reflection.TypeExtensions": "4.1.0-rc3-24026-00",
     "System.Runtime": "4.1.0-rc3-24026-00",
     "System.Runtime.Extensions": "4.1.0-rc3-24026-00",
     "System.Text.RegularExpressions": "4.1.0-rc3-24026-00",
